This is probably the las commit, before I stop working on this project (and possibly one day recreate it, doing good job this time). I repaired sorting bug, added new date to shopping list products, removed some spaces between imports etc. It was fun ^^

+++ app/src/main/java/com/tkkd/mealplanner/Database/Entities/Home.java
@@ -13,7 +13,7 @@ public class Home {
     @PrimaryKey(autoGenerate = true)
     public long id;
 
-    public float quantity;
+    public int quantity;
 
     @ColumnInfo(name = "ingredient_id")
     public long ingId;
